requirements.txt
numpy>=2.0.0
vllm==0.8.3
llama_cpp_python==0.3.8